Updated script..
Added support for selecting non null or output nodes. 
If so an null will be created for the selected node. 

Changed from using root GEO node's name as default to a custom name dialog box. 
If a null or output node is selected, the name of this node is used to pre-populate the dialog box.

Added additional error checks. 
Cleaned up the code a bit